pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-bot"
version = "0.1.0"
description = "Minimal Python customer support bot with CLI + Flask UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flask",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
